requirements.txt
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.4
django-debug-toolbar==2.0
django-filter==2.2.0
django-webpack-loader==0.6.0
djangorestframework==3.10.2
idna==2.8
misago-social-auth-app-django==2.1.0
mysql==0.0.2
mysql-connector-python==8.0.17
mysqlclient==1.4.2.post1
oauthlib==3.1.0
Pillow==6.1.0
postgres==2.2.2
protobuf==3.9.1
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.3.1
PyMySQL==0.9.3
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
selenium==3.141.0
six==1.12.0
social-auth-core==3.2.0
sqlparse==0.3.0
twitter==1.18.0
urllib3==1.25.3
wrapt==1.11.2
